Sort out Identity/RealIdentity confusion Add getLength to Period
