+++ spec/SCL/Repository/Value/NullIdentitySpec.php
@@ -5,6 +5,7 @@ namespace spec\SCL\Repository\Value;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 use SCL\Repository\Exception\NullIdentityException;
+use SCL\Repository\Value\Identity;
 use SCL\Repository\Value\RealIdentity;
 
 class NullIdentitySpec extends ObjectBehavior
@@ -16,7 +17,7 @@ class NullIdentitySpec extends ObjectBehavior
 
     public function it_is_initializable()
     {
-        $this->shouldHaveType(RealIdentity::class);
+        $this->shouldHaveType(Identity::class);
     }
 
     public function it_has_no_id()
@@ -26,6 +27,6 @@ class NullIdentitySpec extends ObjectBehavior
 
     public function it_has_no_value()
     {
-        $this->getValue()->shouldThrow(NullIdentityException::class);
+        $this->shouldThrow(NullIdentityException::class)->duringGetValue();
     }
 }
